backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
Pillow==10.1.0
requests==2.31.0
opencv-python==4.8.1.78
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
werkzeug==3.0.1
PyJWT==2.8.0